Fix: Item, Shop Controller 분리
