+++ src/main/java/com/clovi/item/shop/ShopController.java
@@ -1,0 +1,22 @@
+package com.clovi.item.shop;
+
+import com.clovi.base.dto.response.BaseResponse;
+import com.clovi.base.dto.response.MessageCode;
+import com.clovi.base.dto.response.ProcessStatus;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class ShopController {
+    private final ShopService shopService;
+
+    @GetMapping("/api/v1/shops")
+    public ResponseEntity getShopsV1(){
+        return ResponseEntity.ok(new BaseResponse(shopService.getShops(), HttpStatus.OK.value(), ProcessStatus.SUCCESS, MessageCode.SUCCESS_GET));
+    }
+
+}
